LPS-84026 Add session interface, move open id connect flow state enum to api
